Move method 'getMapVersionByName' to DownloadedMaps.java
